自动把资源绑定AssetBundleNames

+++ Assets/Editor/Build/AssetBundleBuild.cs
@@ -25,42 +25,57 @@ public class AssetBundleBuild{
 		BuildPipeline.BuildAssetBundles(filePath,BuildAssetBundleOptions.None);
 	}
 	
-	[MenuItem("Tools/AssetBundle/AutoSetResourceAssetBundleName")]
-	public static void AutoSetResourceAssetBundleName()
+	[MenuItem("Tools/AssetBundle/AutomaticBinding")]
+	public static void AutomaticBinding()
 	{
-		String filePath = Application.dataPath + "/a.txt.meta";
 		//UI下面的每个目录都分别对应一个bundle
-		
+		foreach(DirectoryInfo child in Directory.CreateDirectory(Application.dataPath + "/Resources/UI").GetDirectories())
+		{
+			AutomaticBindingDirectory(child.FullName,child.Name,false);
+		}
 		//Config打成一个
-		
+		AutomaticBindingDirectory(Application.dataPath + "/Resources/Config","config",false);
 		//Audio下面每个文件打成一个
-		
+		AutomaticBindingDirectory(Application.dataPath + "/Resources/Audio","",true);
 		//Animator 都应该是prefab 每个打成一个
-		
+		AutomaticBindingDirectory(Application.dataPath + "/Resources/Animator","",true,".prefab");
 		//Animation 每个打成一个
-		
-		SetFileAssetBundleName(filePath,"textbubu");
+		AutomaticBindingDirectory(Application.dataPath + "/Resources/Animation","",true,".prefab");
 	}
 	
-	private static void SetDirectoryAssetBundleName(string dirPath,string bundleName)
+	private static void AutomaticBindingDirectory(string directoryPath,string assetBundleName,bool isRespective,string suffix = "*")
 	{
-	
+		DirectoryInfo dir = Directory.CreateDirectory(directoryPath);
+		foreach(DirectoryInfo child in dir.GetDirectories())
+		{
+			AutomaticBindingDirectory(child.FullName,child.Name,isRespective,suffix);
+		}
+		foreach(FileInfo child in dir.GetFiles())
+		{
+			AutomaticBindingFile(child.FullName,isRespective ? child.Name.Substring(0,child.Name.IndexOf(".")) : assetBundleName,suffix);
+		}
 	}
 	
-	private static void SetFileAssetBundleName(String filePath,String bundleName)
+	private static void AutomaticBindingFile(String filePath,String assetBundleName,String suffix)
 	{
-		StreamReader  reader = new StreamReader(filePath);
-		String content = reader.ReadToEnd();
-		reader.Close();
-		
-		content = content.Replace("assetBundleName: \n","assetBundleName: " + bundleName + "\n");
-		
-		File.Delete(filePath);
-		StreamWriter writer = new StreamWriter(filePath,false);
-		writer.Write(content);
-		writer.Flush();
-		writer.Close();
-		
-		Debug.Log("生成成功：" + filePath);
+		if(filePath.LastIndexOf(".meta") >= 0 && (suffix == "*" || filePath.LastIndexOf(suffix) >= 0))
+		{
+			String content = File.ReadAllText(filePath);
+			
+			File.Delete(filePath);
+			
+			int startIndex = content.IndexOf("assetBundleName:");
+			if(startIndex >= 0)
+			{
+				int endIndex = content.IndexOf("\n",startIndex);
+				if(endIndex >= 0)
+				{
+					content = content.Replace(content.Substring(startIndex,endIndex - startIndex),"assetBundleName: " + assetBundleName.ToLower());
+					File.WriteAllText(filePath,content);
+					
+					Debug.Log("生成成功：" + filePath);
+				}
+			}
+		}
 	}
 }
